+++ frontend/src/App.jsx
@@ -10,22 +10,21 @@ import Sidebar from './components/layout/Sidebar';
 
 // Pages
 
-import HomePage from './pages/HomePage';
-import SchedulePage from './pages/SchedulePage';
-import BoothsPage from './pages/BoothsPage';
-import MapSettingsPage from './pages/MapSettingsPage';
-import NoticesPage from './pages/NoticesPage';
-import LostFoundPage from './pages/LostFoundPage';
+import FestivalPage from './pages/FestivalPage';
+import EventPage from './pages/EventPage';
+import PlacePage from './pages/PlacePage';
+import PlaceGeographyPage from './pages/PlaceGeographyPage';
+import AnnouncementPage from './pages/AnnouncementPage';
+import LostItemPage from './pages/LostItemPage';
 import FaqPage from './pages/FaqPage';
 
 // Modals
-import NoticeModal from './components/modals/NoticeModal';
+import AnnouncementModal from './components/modals/AnnouncementModal';
 import LostItemModal from './components/modals/LostItemModal';
 import FaqModal from './components/modals/FaqModal';
-import ScheduleModal from './components/modals/ScheduleModal';
+import EventModal from './components/modals/EventModal';
 import EventDateModal from './components/modals/EventDateModal';
-import BoothModal from './components/modals/BoothModal';
-import CopyLinkModal from './components/modals/CopyLinkModal';
+import PlaceModal from './components/modals/PlaceModal';
 import FestivalInfoModal from './components/modals/FestivalInfoModal';
 import FestivalImagesModal from './components/modals/FestivalImagesModal';
 import PlaceImagesModal from './components/modals/PlaceImagesModal';
@@ -33,7 +32,7 @@ import PlaceEditModal from './components/modals/PlaceEditModal';
 import AddImageModal from './components/modals/AddImageModal';
 import ConfirmModal from './components/common/ConfirmModal';
 import Modal from './components/common/Modal';
-import { NoticeDetailModal } from './components/modals/NoticeModal';
+import { AnnouncementDetailModal } from './components/modals/AnnouncementModal';
 import LineupAddModal from './components/modals/LineupAddModal';
 import LineupEditModal from './components/modals/LineupEditModal';
 import PushNotificationConfirmModal from './components/modals/PushNotificationConfirmModal';
@@ -41,7 +40,6 @@ import PasswordChangeModal from './components/modals/PasswordChangeModal';
 
 // Common Components
 import Toast from './components/common/Toast';
-import FestivalPage from "./pages/FestivalPage.jsx";
 
 function App() {
     const [page, setPage] = useState('home');
@@ -57,20 +55,18 @@ function App() {
     const openModal = (type, props = {}) => setModalState({ type, props });
     const closeModal = () => setModalState({ type: null, props: {} });
 
-
-    
-
     const renderPage = () => {
         switch (page) {
 
-            case 'home': return <HomePage />;
-            case 'schedule': return <SchedulePage />;
-            case 'booths': return <BoothsPage />;
-            case 'map-settings': return <MapSettingsPage />;
-            case 'notices': return <NoticesPage />;
-            case 'lost-found': return <LostFoundPage />;
+            case 'home': return <FestivalPage />;
+            case 'schedule': return <EventPage />;
+            case 'booths': return <PlacePage />;
+            case 'place': return <PlacePage />;
+            case 'map-settings': return <PlaceGeographyPage />;
+            case 'notices': return <AnnouncementPage />;
+            case 'lost-found': return <LostItemPage />;
             case 'faq': return <FaqPage />;
-            default: return <HomePage />;
+            default: return <FestivalPage />;
         }
     };
 
@@ -78,14 +74,13 @@ function App() {
         const { type, props } = modalState;
         const allProps = { ...props, onClose: closeModal, showToast, openModal };
         switch (type) {
-            case 'notice': return <NoticeModal {...allProps} />;
-            case 'notice-detail': return <NoticeDetailModal {...allProps} />;
+            case 'notice': return <AnnouncementModal {...allProps} />;
+            case 'notice-detail': return <AnnouncementDetailModal {...allProps} />;
             case 'lostItem': return <LostItemModal {...allProps} />;
             case 'faq': return <FaqModal {...allProps} />;
-            case 'schedule': return <ScheduleModal {...allProps} />;
+            case 'schedule': return <EventModal {...allProps} />;
             case 'datePrompt': return <EventDateModal {...allProps} />;
-            case 'booth': return <BoothModal {...allProps} />;
-            case 'copyLink': return <CopyLinkModal {...allProps} />;
+            case 'booth': return <PlaceModal {...allProps} />;
             case 'festival-info': return <FestivalInfoModal isOpen={true} {...allProps} />;
             case 'festival-images': return <FestivalImagesModal isOpen={true} {...allProps} />;
             case 'festivalImages': return <FestivalImagesModal isOpen={true} {...allProps} />;
@@ -93,7 +88,6 @@ function App() {
             case 'placeEdit': return <PlaceEditModal {...allProps} />;
             case 'confirm': return <ConfirmModal {...allProps} onConfirm={() => { props.onConfirm(); closeModal(); }} onCancel={closeModal} />;
             case 'image': return <Modal isOpen={true} onClose={closeModal} maxWidth="max-w-4xl"><div className="relative"><img src={props.src} className="max-w-full max-h-[80vh] rounded-lg mx-auto" alt="상세 이미지" /><button onClick={closeModal} className="absolute top-2 right-2 text-white text-3xl bg-black bg-opacity-50 rounded-full w-8 h-8 flex items-center justify-center">&times;</button></div></Modal>;
-            case 'festival': return <FestivalPage {...allProps} />;
             case 'lineup-add': return <LineupAddModal isOpen={true} {...allProps} />;
             case 'lineup-edit': return <LineupEditModal isOpen={true} {...allProps} />;
             case 'add-image': return <AddImageModal isOpen={true} {...allProps} />;
@@ -110,7 +104,6 @@ function App() {
                 <ModalContext.Provider value={{ openModal, closeModal, showToast }}>
                     <div className="bg-gray-100 text-gray-800 flex h-screen">
 
-                        {localStorage.getItem('festivalId') === null ? <FestivalPage/> : <></>}
                         <Sidebar open={sidebarOpen} setOpen={setSidebarOpen} />
                         <main className="flex-1 p-8 overflow-y-auto">
                             {renderPage()}
